creation clientservice

+++ public/deleteclient.js
@@ -2,19 +2,31 @@ const btns = document.querySelectorAll(".btn-delete");
 
 btns.forEach(btn => {
   btn.addEventListener("click", () => {
-    const clientChoice = confirm("Êtes-vous sûr de vouloir supprimer cet utilisateur ?")
-    if (!clientChoice) return; // retourn et n'execute pas le fetch
     const clientId = btn.id;
+// verification si l'ID est valid (pour eviter que des valeurs inatendue soient passées à la fonction)
+    
+if(!clientId || isNaN(clientId)){
+  console.log("ID client invalide");
+  return;
+}
+const clientChoice = confirm("Êtes-vous sûr de vouloir supprimer cet utilisateur ?")
+    if (!clientChoice) return; // retourn et n'execute pas le fetch
     supprimerClient(clientId)
   })
 });
 
-
 // Fonction pour afficher le popup
 function afficherPopup(message) {
   const popup = document.getElementById('popup-success');
-  document.getElementById('popup-message').textContent = message; // Mettre à jour le message
-  popup.style.display = 'block'; // Afficher le popup
+  const safeMessage = document.createTextNode(message)
+  const messageContainer = document.getElementById(`popup-message`)
+  //document.getElementById('popup-message').textContent = message; // Mettre à jour le message
+
+  //effacer les anciens messages et ajouter le nouveau
+  messageContainer.innerHTML``;
+  messageContainer.appendChild(safeMessage);
+
+  popup.style.display = 'none'; // Afficher le popup
   setTimeout(() => {
     fermerPopup();
   }, 3000)
@@ -58,6 +70,15 @@ function supprimerClient(id) {
       console.error('Erreur lors de la suppression du client:', error);
     });
 }
+
+/* Ajout de sécurité a 2 endroits:
+
+1 vérification si l'id est valide
+2 Sécurisation du popup (du message envoyer)
+
+
+*/
+
